re-organize common
