--- libs/common/src/enums.ts
@@ -1,7 +1,0 @@
-export enum EventType {
-  PUSH = 'PUSH',
-  PULL_REQUEST_OPENED = 'PULL_REQUEST_OPENED',
-  PULL_REQUEST_MERGED = 'PULL_REQUEST_MERGED',
-  PULL_REQUEST_REVIEWED = 'PULL_REQUEST_REVIEWED',
-  WORKFLOW_RUN_COMPLETED = 'WORKFLOW_RUN_COMPLETED',
-}
