+++ libs/common/src/interfaces/pagination.interface.ts
@@ -1,10 +1,3 @@
-export interface CreateUserInput {
-  uid?: string;
-  email: string | null;
-  displayName: string | null;
-  photoUrl: string | null;
-  // githubUsername?: string;
-}
 
 export interface PaginationParams {
   limit?: number;
