+++ libs/common/src/models/common.model.ts
@@ -1,3 +1,9 @@
+export type ConnectedAccountProvider =
+  | 'github'
+  | 'strava'
+  | 'azure'
+  | 'bitbucket';
+
 export interface BaseDocument {
   id: string;
   createdAt: string; // ISO string
