+++ libs/common/src/models/index.ts
@@ -1,3 +1,5 @@
 export * from './activity.model';
 export * from './common.model';
+export * from './event.model';
 export * from './user.model';
+
