+++ libs/common/src/models/user.model.ts
@@ -1,5 +1,13 @@
 import { BaseDocument } from './common.model';
 
+export interface CreateUserInput {
+  uid?: string;
+  email: string | null;
+  displayName: string | null;
+  photoUrl: string | null;
+  // githubUsername?: string;
+}
+
 // User Document
 export interface User extends BaseDocument {
   email: string;
