--- libs/common/src/types.ts
@@ -1,5 +1,0 @@
-export type ConnectedAccountProvider =
-  | 'github'
-  | 'strava'
-  | 'azure'
-  | 'bitbucket';
